requirements.txt
fastapi
openai
python-dotenv
psycopg2
joblib